move result module
domain to core
