+++ domain/src/main/java/com/rmakiyama/spatz/domain/repository/AuthRepository.kt
@@ -1,7 +1,6 @@
 package com.rmakiyama.spatz.domain.repository
 
 import com.rmakiyama.spatz.domain.model.auth.AuthUser
-import com.rmakiyama.spatz.domain.result.Result
 import kotlinx.coroutines.flow.Flow
 
 interface AuthRepository {
