+++ features/home/src/main/java/com/rmakiyama/spatz/home/HomeFragment.kt
@@ -16,7 +16,7 @@ import com.rmakiyama.spatz.core.destination.ScreenDestination
 import com.rmakiyama.spatz.core.extension.KEY_LOGIN_SUCCESSFUL
 import com.rmakiyama.spatz.domain.model.auth.AuthUser
 import com.rmakiyama.spatz.domain.model.tweet.Tweet
-import com.rmakiyama.spatz.domain.result.Result
+import com.rmakiyama.spatz.core.result.Result
 import com.rmakiyama.spatz.home.databinding.FragmentHomeBinding
 import com.rmakiyama.spatz.home.item.TweetItem
 import com.xwray.groupie.GroupAdapter
