+++ usecase/src/main/java/com/rmakiyama/spatz/usecase/SuspendUseCase.kt
@@ -1,6 +1,6 @@
 package com.rmakiyama.spatz.usecase
 
-import com.rmakiyama.spatz.domain.result.Result
+import com.rmakiyama.spatz.core.result.Result
 
 abstract class SuspendUseCase<in Command, out Type> {
