+++ usecase/src/main/java/com/rmakiyama/spatz/usecase/auth/LoadAuthUserUseCase.kt
@@ -2,7 +2,7 @@ package com.rmakiyama.spatz.usecase.auth
 
 import com.rmakiyama.spatz.domain.model.auth.AuthUser
 import com.rmakiyama.spatz.domain.repository.AuthRepository
-import com.rmakiyama.spatz.domain.result.Result
+import com.rmakiyama.spatz.core.result.Result
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
